// frontend/src/types/AllianceNftMinter.d.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@0.35.3.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export type Addr = string;
export type Timestamp = Uint64;
export type Uint64 = string;
export interface InstantiateMsg {
  dao_treasury_address?: Addr | null;
  mint_end_time: Timestamp;
  mint_start_time: Timestamp;
  nft_collection_code_id: number;
}
export type ExecuteMsg =
  | {
      append_nft_metadata: {
        [k: string]: MinterExtension;
      };
    }
  | {
      mint: object;
    }
  | {
      send_to_dao: number;
    }
  | {
      change_dao_treasury_address: string;
    }
  | {
      change_owner: string;
    };
export interface MinterExtension {
  extension: Metadata;
  token_id: string;
}
export interface Metadata {
  animation_url?: string | null;
  attributes?: Trait[] | null;
  background_color?: string | null;
  description?: string | null;
  external_url?: string | null;
  image?: string | null;
  image_data?: string | null;
  name?: string | null;
  youtube_url?: string | null;
}
export interface Trait {
  display_type?: string | null;
  trait_type: string;
  value: string;
}
export type QueryMsg =
  | {
      config: object;
    }
  | {
      stats: object;
    }
  | {
      nft_data: string;
    };
export interface MinterConfig {
  dao_treasury_address?: Addr | null;
  mint_end_time: Timestamp;
  mint_start_time: Timestamp;
  nft_collection_address?: Addr | null;
  owner: Addr;
}
export interface MinterStats {
  available_nfts: number;
  minted_nfts: number;
}